separate search bar into it's own
